Low: fencing: Improvements to stonith-test using fence_true and fence_false.
This patch re-works how stonith test works by using
the fence_true and fence_false agents to sanity check
the client api.  When stonith-test executable is run
on the same machine as the stonithd daemon, stonith-test
will return 0 on success success and -1 on test failure.
The -V option can be used to view verbose log output
to the cli.