High: cib: Prevent old digest from making it into cib broadcast.
The fix_cib_diff() function was recently moved into cib_config_changed().
This caused the digest of an incomplete cib update to be calculated and placed
on the diff result.  In the cib_perform_op() function cib_config_changed is called
to determine if certain fields in the new cib need to be incremented.  Because of this
the digest is calculated and stored in the diff before all the changes are made
to the new cib result.  The fix_cib_diff() fuction is already called later on in
cib_perform_op() at the correct location, so simply removing fix_cib_diff from
cib_config_changed() fixes most of the issue.